move some testing helper functions to a new `optype.inspect` module
